+++ perception/compare_map_segmentation/CMakeLists.txt
@@ -22,12 +22,12 @@ include_directories(
 )
 
 add_library(${PROJECT_NAME} SHARED
-  src/distance_based_compare_map_filter_nodelet.cpp
-  src/voxel_based_approximate_compare_map_filter_nodelet.cpp
-  src/voxel_based_compare_map_filter_nodelet.cpp
-  src/voxel_distance_based_compare_map_filter_nodelet.cpp
-  src/compare_elevation_map_filter_node.cpp
-  src/voxel_grid_map_loader.cpp
+  src/distance_based_compare_map_filter/node.cpp
+  src/voxel_based_approximate_compare_map_filter/node.cpp
+  src/voxel_based_compare_map_filter/node.cpp
+  src/voxel_distance_based_compare_map_filter/node.cpp
+  src/compare_elevation_map_filter/node.cpp
+  src/voxel_grid_map_loader/voxel_grid_map_loader.cpp
 )
 
 target_link_libraries(${PROJECT_NAME}
