+++ perception/compare_map_segmentation/src/voxel_based_approximate_compare_map_filter/node.hpp
@@ -12,11 +12,11 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#ifndef COMPARE_MAP_SEGMENTATION__VOXEL_BASED_APPROXIMATE_COMPARE_MAP_FILTER_NODELET_HPP_  // NOLINT
-#define COMPARE_MAP_SEGMENTATION__VOXEL_BASED_APPROXIMATE_COMPARE_MAP_FILTER_NODELET_HPP_  // NOLINT
+#ifndef VOXEL_BASED_APPROXIMATE_COMPARE_MAP_FILTER__NODE_HPP_  // NOLINT
+#define VOXEL_BASED_APPROXIMATE_COMPARE_MAP_FILTER__NODE_HPP_  // NOLINT
 
+#include "../voxel_grid_map_loader/voxel_grid_map_loader.hpp"
 #include "pointcloud_preprocessor/filter.hpp"
-#include "voxel_grid_map_loader.hpp"
 
 #include <pcl/filters/voxel_grid.h>
 #include <pcl/search/pcl_search.h>
@@ -70,8 +70,8 @@ public:
   PCL_MAKE_ALIGNED_OPERATOR_NEW
   explicit VoxelBasedApproximateCompareMapFilterComponent(const rclcpp::NodeOptions & options);
 };
-}  // namespace compare_map_segmentation
+}  // namespace autoware::compare_map_segmentation
 
 // clang-format off
-#endif  // COMPARE_MAP_SEGMENTATION__VOXEL_BASED_APPROXIMATE_COMPARE_MAP_FILTER_NODELET_HPP_  // NOLINT
+#endif  // VOXEL_BASED_APPROXIMATE_COMPARE_MAP_FILTER__NODE_HPP_  // NOLINT
 // clang-format on
